add test fixtures
