+++ eventsourcing_helpers/messagebus/__init__.py
@@ -6,7 +6,7 @@ from eventsourcing_helpers.utils import import_backend
 
 BACKENDS = {
     "kafka_avro": "eventsourcing_helpers.messagebus.backends.kafka.KafkaAvroBackend",
-    "mock": "eventsourcing_helpers.messagebus.backends.mock.MockBackend",
+    "mock": "eventsourcing_helpers.messagebus.backends.mock.backend.MockBackend",
 }
 
 logger = structlog.get_logger(__name__)
