+++ setup.py
@@ -2,13 +2,16 @@ from setuptools import find_packages, setup
 
 setup(
     name="eventsourcing-helpers",
-    version="1.0.2",
+    version="1.1.0.rc0",
     description="Helpers for practicing the Event sourcing pattern",
     url="https://github.com/fyndiq/eventsourcing_helpers",
     author="Fyndiq AB",
     author_email="support@fyndiq.com",
     license="MIT",
     packages=find_packages(),
+    entry_points={
+        "pytest11": ["eventsourcing_helpers = eventsourcing_helpers.messagebus.backends.mock.test"]
+    },
     install_requires=[
         "structlog>=17.2.0",
         "jsonpickle>=0.9.6",
